Prepend the name of the mother volume to the PMT "tubeTag".
In order for hits from PMT volumes to be associated with the
appropriate entries in the event data structure, a unique name is
derived from every PMTs location in the geometry.  Before these had
the form:

tubeLocationmap[WCPMTGlass:0:0:0:0:0:5]= 6

This leads to collisions if you have a different PMT array with a
similar hierarchy of replica numbering.  Now the tubeTag has been
expanded to include the name of the mother volume, which should
distinguish PMTs in the barrel from things like PMTs in a veto.
Example:

tubeLocationmap[WCBarrelCell:WCPMTGlass:0:0:0:0:0:5]= 6

Comments have also been added to warn people to keep these two files
in sync when changing the tubeTag in the future.