Checkpoint of new LBNE digitizer simulation + sample tests.
The digitizer simulation (including the application of arbitrary PMT
charge and time distributions) is basically finished with unit tests
covering most of the new code functionality.  (The included wcsimtest
doesn't actually test the digitizer code.  It is just an example.)

I've also added a "shim" class to make it a little easier to swap the
WCSimLBNEDigitizer implementation in place of the WCSimWCDigitizer
implementation in the end of event action.

Until the configuration file interface is complete, a control variable
at the top of WCSimEventAction is used to select between the two
models at compile time.  This is just for rough testing now and will
be removed before this branch is finished.  Time and charge
distributions are made-up Gaussians until real distributions are
loaded via the configuration files.