Comments, function name change and moving xs_test
Added comments to reference all equations with which part of the
Strummia and Vissani paper they came from. Moves the xs_test declaration
back into the while loop and a comment explaining this. Changed the name
of the "Current-Current structure" function to MatrixElement to better
reflect the calculation. It is calculating the matrix element which has
a current-current structure.